refactor: move backfill code to temp module
