+++ backend/mlarchive/archive/storage.py
@@ -6,14 +6,10 @@ import json
 from contextlib import contextmanager
 from storages.backends.s3 import S3Storage
 
-from django.conf import settings
 from django.core.files.base import File
-from django.db.models.functions import Length
 from django.utils import timezone
-from django.utils.module_loading import import_string
 
 from mlarchive.archive.models import StoredObject
-from mlarchive.blobdb.models import Blob
 from mlarchive.blobdb.storage import BlobdbStorage, MetadataFile
 
 import logging
@@ -202,61 +198,3 @@ class StoredObjectBlobdbStorage(BlobdbStorage):
     def delete(self, name):
         self._delete_stored_object(name)
         super().delete(name)
-
-
-def tracked_buckets():
-    """Return the blobdb bucket names whose storage records StoredObject rows"""
-    buckets = []
-    for config in settings.STORAGES.values():
-        backend = import_string(config['BACKEND'])
-        if issubclass(backend, StoredObjectBlobdbStorage):
-            buckets.append(config['OPTIONS']['bucket_name'])
-    return buckets
-
-
-def backfill_stored_objects(start_after_pk=0, batch_size=5000):
-    """Index one batch of pre-existing blobs as StoredObject rows"""
-
-    rows = list(
-        Blob.objects
-        .filter(pk__gt=start_after_pk, bucket__in=tracked_buckets())
-        .order_by('pk')
-        .annotate(object_size=Length('content'))
-        .values_list('pk', 'bucket', 'name', 'checksum', 'object_size', 'modified')
-        [:batch_size]
-    )
-    if not rows:
-        return None
-
-    existing = set()
-    for bucket in {row[1] for row in rows}:
-        names = [row[2] for row in rows if row[1] == bucket]
-        existing.update(
-            StoredObject.objects
-            .filter(store=bucket, name__in=names)
-            .values_list('store', 'name')
-        )
-
-    records = [
-        StoredObject(
-            store=bucket,
-            name=name,
-            sha384=checksum,
-            len=object_size,
-            store_created=modified,
-            created=modified,
-            modified=modified,
-        )
-        for _, bucket, name, checksum, object_size, modified in rows
-        if (bucket, name) not in existing
-    ]
-    # ignore_conflicts covers a row written by the storage between the existence
-    # check and the insert; the explicit batch_size keeps each INSERT within the
-    # Postgres parameter limit regardless of the caller's batch size.
-    StoredObject.objects.bulk_create(records, batch_size=1000, ignore_conflicts=True)
-
-    return {
-        'last_pk': rows[-1][0],
-        'created': len(records),
-        'skipped': len(rows) - len(records),
-    }
